+++ MagRetentionReload/ModInfo.cs
@@ -1,0 +1,11 @@
+namespace Tosox.MagRetentionReload
+{
+    internal static class ModInfo
+    {
+        internal const string Guid = "de.tosox.magretentionreload";
+        internal const string Name = "Mag Retention Reload";
+        internal const string Author = "Tosox";
+        internal const string Source = "https://github.com/Tosox/SPT-MagRetentionReload";
+        internal const string Version = "1.3.0";
+    }
+}